Quite a few changes:
Implementation of geography as a dimension (role: Spatial, isKey: True)

Implemented a list of known geography variables on Cantabular which will now default to geography. This is accessible inside of config.json

Implemented isKey so now it will make the geography dimension key and if there is not a geography dimension then it will make first query variable the key

Reworked the updating dataset logic. It will now only let you update the dataset if the non-geography dimensions are equal to each other.

Added a requirements.txt file which contains all the dependencies which can be used by 'pip install -r requirements.txt'

TO DO: Change the PUT observations to a POST if its updating. Ask Spencer about how to handle geography dimension if it is different when updating.