Improved logging, tidied types and docstrings, addition of debug flag
The logging functionality of the program has been improved such that now only relevant events are logged to the file, including any exceptions that halt the program.  The verbose flag has been altered so as to not print out literally everything, and the debug flag was added to do just that.